==> 1-get-data.ipynb <==
import requests
from bs4 import BeautifulSoup
import os
import pandas as pd
import re

soup = BeautifulSoup(requests.get("https://southpark.fandom.com/wiki/Portal:Scripts").text)

BASE_URL = "https://southpark.fandom.com"
season_urls = soup.find('div', id='gallery-0').find_all("a")
urls = {}
for tag in season_urls:
    urls[tag["href"]] = None
season_urls = list(map(lambda key: BASE_URL + key, urls.keys()))
season_urls

def get_episode_urls(season_url: url):
    season_data = BeautifulSoup(requests.get(season_url).text)
    episode_overview = season_data.find('div', id='gallery-0').find_all("a")

    episode_urls = {}
    for episode in episode_overview:
        episode_urls[episode["href"]] = None
    return list(map(lambda key: BASE_URL + key, episode_urls.keys()))

def get_episode_script(episode_url: str, season: int, episode: int):
    """ Given the full url (base + episode url), scrape the script and return title + lines """
    episode_data = requests.get(episode_url).text
    script_rows = BeautifulSoup(episode_data).find_all("table", class_="headerscontent")[-1].find_all("tr", style="")
    # get title
    episode_title = script_rows[0].find("td").text.strip()
    print(f"Episode title: {episode_title}")
    # get script
    script_rows = script_rows[1:]
    lines = []
    for row in script_rows:
        line = row.find_all("td")
        actor = line[0].text.strip()
        dialogue = line[1].text.strip()
        lines.append([season, episode, actor, dialogue])
    return episode_title, lines

def slugify(text: str):
    # Replace non-alphanumeric characters (except hyphens and underscores) with hyphens
    slug = re.sub(r'[^a-zA-Z0-9-_]', '-', text)
    
    # Replace consecutive hyphens with a single hyphen
    slug = re.sub(r'-+', '-', slug)
    
    # Remove leading and trailing hyphens
    slug = slug.strip('-')
    
    # Convert to lowercase
    slug = slug.lower()
    
    return slug

SEASON = 1
for season_url in season_urls:
    episode_urls = get_episode_urls(season_url)
    EPISODE = 1
    for episode_url in episode_urls:
        episode_title, episode_lines = get_episode_script(episode_url, SEASON, EPISODE)
        episode_title = slugify(episode_title)
        df = pd.DataFrame(episode_lines)
        dir = os.path.join(".", "episodes_csv", f"s{SEASON}")
        if not os.path.exists(dir):
            os.makedirs(dir)
        # Save the DataFrame to a CSV file with a custom delimiter (semicolon in this case)
        df.to_csv(os.path.join(dir, f"e{EPISODE}_{episode_title}.csv"), sep=';', index=False, header=False)
        EPISODE += 1
    SEASON += 1

### Get synopsis of each episode

f = open('episode_overview_edited.html', 'r')
lines = f.read()
f.close()
soup = BeautifulSoup(lines)

episode_title_rows = soup.find_all('tr', attrs={'style': 'text-align:center;'})
synopsis_dict = {}

for episode_row in episode_title_rows:
    fields = episode_row.find_all("td")
    ep_name = fields[1].text.strip()
    ep_code = fields[3].text.lstrip().rstrip()
    synopsis = episode_row.next_sibling.next_sibling.find("p").text.lstrip().rstrip().replace("\n", "").replace("\'", "`")
    synopsis_dict[ep_code] = synopsis

f = open("synopsis_mappings.txt", "w")
f.write(str(synopsis_dict))
f.close()
