==> 3-tune-agents.ipynb <==
import os
from openai import OpenAI

# get api key from file
API_KEY = ""
with open("key.secret", "r") as f:
    API_KEY = f.read()

client = OpenAI(api_key=API_KEY)

### Upload tuning files

tuning_path = os.path.join(".", "tuning_data")
tuning_files = ["butters_it.jsonl", "cartman_it.jsonl", "kyle_it.jsonl", "manager_it.jsonl", "stan_it.jsonl"]

uploaded_files = []

for file in tuning_files:
    path = os.path.join(tuning_path, file)
    uploaded_file = client.files.create(
      file=open(path, "rb"),
      purpose="fine-tune",
    )
    uploaded_files.append(uploaded_file)

uploaded_files

### Upload eval (validation) files

eval_files = ["butters_eval.jsonl", "cartman_eval.jsonl", "kyle_eval.jsonl", "manager_eval.jsonl", "stan_eval.jsonl"]

uploaded_files_eval = []

for file in eval_files:
    path = os.path.join(tuning_path, file)
    uploaded_file = client.files.create(
      file=open(path, "rb"),
      purpose="fine-tune",
    )
    uploaded_files_eval.append(uploaded_file)

uploaded_files_eval

### Do the tuning

def find_file(l: list, name: str):
    for file in l:
        if name in file.filename:
            return file

def start_tuning_job(model_name, it_file, eval_file):
    client.fine_tuning.jobs.create(
        training_file=it_file.id, 
        model="gpt-3.5-turbo",
        suffix=model_name,
        validation_file=eval_file.id,
        hyperparameters={
            "n_epochs":3
        }
    )

model_name = "butters"
it_file = find_file(uploaded_files, model_name)
eval_file = find_file(uploaded_files_eval, model_name)
        
start_tuning_job(model_name, it_file, eval_file)

model_name = "cartman"
it_file = find_file(uploaded_files, model_name)
eval_file = find_file(uploaded_files_eval, model_name)
        
start_tuning_job(model_name, it_file, eval_file)

model_name = "kyle"
it_file = find_file(uploaded_files, model_name)
eval_file = find_file(uploaded_files_eval, model_name)
        
start_tuning_job(model_name, it_file, eval_file)

model_name = "stan"
it_file = find_file(uploaded_files, model_name)
eval_file = find_file(uploaded_files_eval, model_name)
        
start_tuning_job(model_name, it_file, eval_file)

model_name = "manager"
it_file = find_file(uploaded_files, model_name)
eval_file = find_file(uploaded_files_eval, model_name)
        
start_tuning_job(model_name, it_file, eval_file)
